Updated dependencies. Made ICS.s.lols set to clear.
Per 20.3.5 of KEA64RM Rev. 2 (pg 276) LOLS register is writable as "w1c", meaning write 1 to clear field.
Per CMSIS-SVD v1.3.6, the equivalent functionality is the modifiedWriteValues parameter with a value of "oneToSet".

There are likely other fields that should set to clear. Need to search through manual for fields marked "w1c" and compare to the svd file.